+++ src/tui/tabbed.rs
@@ -9,6 +9,9 @@ use tui::messaging::Timestamp;
 use tui::MsgTarget;
 use tui::widget::WidgetRet;
 
+use notifier::Notifier;
+use notifier::NotifyFor;
+
 const LEFT_ARROW: char = '<';
 const RIGHT_ARROW: char = '>';
 
@@ -29,6 +32,7 @@ struct Tab {
     style: TabStyle,
     /// Alt-character to use to switch to this tab.
     switch: Option<char>,
+    notifier: Notifier,
 }
 
 // NOTE: Keep the variants sorted in increasing significance, to avoid updating
@@ -184,7 +188,7 @@ impl Tabbed {
         }
     }
 
-    fn new_tab(&mut self, idx: usize, src: MsgSource, status: bool) {
+    fn new_tab(&mut self, idx: usize, src: MsgSource, status: bool, notify_for: NotifyFor) {
         use std::collections::HashMap;
 
         let mut switch_keys: HashMap<char, i8> = HashMap::with_capacity(self.tabs.len());
@@ -223,6 +227,7 @@ impl Tabbed {
                 src,
                 style: TabStyle::Normal,
                 switch,
+                notifier: Notifier::init(notify_for)
             },
         );
     }
@@ -237,7 +242,8 @@ impl Tabbed {
                     MsgSource::Serv {
                         serv_name: serv_name.to_owned(),
                     },
-                    true
+                    true,
+                    NotifyFor::Mentions
                 );
                 Some(tab_idx)
             }
@@ -268,10 +274,12 @@ impl Tabbed {
                     }
                     Some(serv_tab_idx) => {
                         let mut status_val: bool = true;
+                        let mut notify_for: NotifyFor = NotifyFor::Mentions;
                         for tab in self.tabs.iter() {
                             if let MsgSource::Serv{ serv_name: ref serv_name_ } = tab.src {
                                 if serv_name == serv_name_ {
                                     status_val = tab.widget.get_ignore_state();
+                                    notify_for = tab.notifier.notify_for;
                                     break
                                 }
                             }
@@ -283,7 +291,8 @@ impl Tabbed {
                                 serv_name: serv_name.to_owned(),
                                 chan_name: chan_name.to_owned(),
                             },
-                            status_val
+                            status_val,
+                            notify_for
                         );
                         if self.active_idx >= tab_idx {
                             self.next_tab();
@@ -324,7 +333,8 @@ impl Tabbed {
                                 serv_name: serv_name.to_owned(),
                                 nick: nick.to_owned(),
                             },
-                            true
+                            true,
+                            NotifyFor::Messages
                         );
                         if let Some(nick) = self.tabs[tab_idx].widget.get_nick().map(str::to_owned) {
                             self.tabs[tab_idx + 1].widget.set_nick(nick);
@@ -909,6 +919,12 @@ impl Tabbed {
         });
     }
 
+    pub fn add_client_notify_msg(&mut self, msg: &str, target: &MsgTarget) {
+        self.apply_to_target(target, &|tab: &mut Tab, _| {
+            tab.widget.add_client_notify_msg(msg);
+        });
+    }
+
     pub fn add_client_msg(&mut self, msg: &str, target: &MsgTarget) {
         self.apply_to_target(target, &|tab: &mut Tab, _| {
             tab.widget.add_client_msg(msg);
@@ -925,6 +941,10 @@ impl Tabbed {
     ) {
         self.apply_to_target(target, &|tab: &mut Tab, _| {
             tab.widget.add_privmsg(sender, msg, ts, false, ctcp_action);
+            let nick = tab.widget.get_nick();
+            if let Some(nick_) = nick {
+                tab.notifier.notify_privmsg(sender, msg, target, nick_, false);
+            }
         });
     }
 
@@ -938,6 +958,10 @@ impl Tabbed {
     ) {
         self.apply_to_target(target, &|tab: &mut Tab, _| {
             tab.widget.add_privmsg(sender, msg, ts, true, ctcp_action);
+            let nick = tab.widget.get_nick();
+            if let Some(nick_) = nick {
+                tab.notifier.notify_privmsg(sender, msg, target, nick_, true);
+            }
         });
     }
 
@@ -1042,6 +1066,28 @@ impl Tabbed {
         }
         false
     }
+
+    pub fn notify(&mut self, notify_for: NotifyFor, target: &MsgTarget){
+        self.apply_to_target(target, &|tab: &mut Tab, _| {
+            tab.notifier.set_notify_for(notify_for);
+        });
+    }
+
+    pub fn show_notify_mode(&mut self, target: &MsgTarget){
+        self.apply_to_target(target, &|tab: &mut Tab, _| {
+            let notify_for = tab.notifier.notify_for;
+            if notify_for == NotifyFor::Off {
+                tab.widget.add_client_notify_msg("Notifications are off");
+            }
+            else if notify_for == NotifyFor::Mentions {
+                tab.widget.add_client_notify_msg("Notifications enabled for mentions");
+            }
+            else if notify_for == NotifyFor::Messages {
+                tab.widget.add_client_notify_msg("Notifications enabled for all messages");
+            }
+        });
+    }
+
 
     ////////////////////////////////////////////////////////////////////////////
     // Helpers